dqn/capture/input_mac.py
import Quartz

def press_key(vkey):
    Quartz.CGEventPost(Quartz.kCGHIDEventTap, Quartz.CGEventCreateKeyboardEvent(None, vkey, True))

def release_key(vkey):
    Quartz.CGEventPost(Quartz.kCGHIDEventTap, Quartz.CGEventCreateKeyboardEvent(None, vkey, False))
