requirements.txt
mss
Pillow
pynput
appscript
pyobjc-framework-Quartz
pywin32
msgpack
pygame